Tolls/requirements.txt.txt
pyqrcode
pypng
captcha